requirements.txt
asgiref==3.8.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
backcall==0.2.0
build==0.10.0
CacheControl==0.12.14
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
constantly==15.1.0
contourpy==1.0.7
crashtest==0.4.1
cryptography==39.0.1
cycler==0.11.0
daphne==4.0.0
dash==2.8.1
dash-bootstrap-components==1.3.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
debugpy==1.6.6
decorator==5.1.1
distlib==0.3.7
Django==4.2
django-plotly-dash==2.1.3
django-redis==5.2.0
djangorestframework==3.15.2
dpd-components==0.1.0
dulwich==0.21.5
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.12.2
Flask==2.2.2
fonttools==4.38.0
greenlet==2.0.2
gunicorn==20.1.0
html5lib==1.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
incremental==22.10.0
installer==0.7.0
ipykernel==6.21.2
ipython==8.10.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.0.2
jupyter_core==5.2.0
keyring==23.13.1
kiwisolver==1.4.4
lightgbm==3.3.5
lockfile==0.12.2
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
more-itertools==10.0.0
msgpack==1.0.4
nest-asyncio==1.5.6
numpy==1.24.2
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.4.0
pkginfo==1.9.6
platformdirs==3.10.0
plotly==5.13.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
PyMySQL==1.0.3
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
pywin32-ctypes==0.2.2
pyzmq==25.0.0
rapidfuzz==2.15.1
redis==4.5.1
referencing==0.30.0
requests==2.31.0
requests-toolbelt==1.0.0
rpds-py==0.9.2
scikit-learn==1.2.1
scipy==1.10.0
service-identity==21.1.0
shellingham==1.5.0.post1
six==1.16.0
sklearn==0.0.post1
SQLAlchemy==2.0.15
sqlparse==0.5.0
stack-data==0.6.2
tenacity==8.2.1
threadpoolctl==3.1.0
tomli==2.0.1
tomlkit==0.12.1
tornado==6.2
traitlets==5.9.0
trove-classifiers==2023.7.6
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.16
virtualenv==20.24.2
waitress==2.1.2
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.3
xgboost==1.7.3
zipp==3.13.0
zope.interface==5.5.2
